% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blow.R
\name{add_blow}
\alias{add_blow}
\title{Bind the weighted log odds to a tidy dataset}
\source{
\url{https://doi.org/10.1093/pan/mpn018}
}
\usage{
add_blow(
  tbl,
  group,
  feature,
  n,
  topic = NULL,
  .prior = c("informed", "empirical", "uninformed"),
  .k_prior = 0.1,
  .alpha_prior = 1,
  .compare = c("dataset", "groups"),
  .complete = FALSE,
  .unweighted = TRUE,
  .variance = TRUE,
  .odds = FALSE,
  .prob = FALSE
)
}
\arguments{
\item{tbl}{A tidy dataset with one row per feature and set}

\item{group}{Column of groups between which to compare features, such as
documents for text data}

\item{feature}{Column of features for identifying differences, such as words or
bigrams with text data}

\item{n}{Column containing feature-set counts}

\item{topic}{(Optional) topic to compare groups within}

\item{.prior}{Whether prior should be based on g-prior from empirical Bayes (informed),
total frequency count (empirical), or uninformed with set alpha (uninformed)}

\item{.k_prior}{Penalty term for informed prior}

\item{.alpha_prior}{Frequency of each feature for uninformed prior}

\item{.compare}{Whether to compare group-feature to entire dataset or
against all other groups}

\item{.complete}{Whether to complete all topic-group-feature combinations}

\item{.unweighted}{Whether to include point estimate log odds}

\item{.variance}{Whether to include variance of feature}

\item{.odds}{Whether to include odds of seeing feature within group}

\item{.prob}{Whether to include probability for feature within group}
}
\description{
Calculate and bind the log odds ratio, weighted by a Dirichlet
prior, of a tidy dataset to the dataset itself. The weighted log odds ratio
is added as a column. This functions supports non-standard evaluation through
the tidyeval framework.
}
\details{
The arguments \code{group}, \code{feature}, \code{n}, and \code{topic}
are passed by expression and support \link[rlang]{quasiquotation};
you can unquote strings and symbols. Grouping is preserved but ignored.

The dataset must have exactly one row per topic-group-feature combination for
this calculation to succeed. Read Monroe, Colaresi, and Quinn (2017) for
more on the weighted log odds ratio.
}
