% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_normal.R
\name{add_normal_post}
\alias{add_normal_post}
\title{Add normal posterior to data frame}
\usage{
add_normal_post(df, x, ..., .mean0 = NULL, .sd0 = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{x}{Vector}

\item{...}{Grouping (optional)}

\item{.mean0}{Prior mean, if null then calculated empirically}

\item{.sd0}{Prior SD, if null then calculated empirically}
}
\description{
Add normal posterior to data frame
}
