% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prop.R
\name{add_prop}
\alias{add_prop}
\title{Add proportion to counts}
\usage{
add_prop(
  df,
  n,
  ...,
  sort = TRUE,
  round = FALSE,
  denom = "sum",
  pct_formatting = FALSE,
  cum_prop = FALSE
)
}
\arguments{
\item{df}{a tibble with counts}

\item{n}{column of counts}

\item{...}{variables to group by}

\item{sort}{if TRUE will sort output in descending order of n}

\item{round}{if TRUE will round to nearest percent}

\item{denom}{denominator for calculating percent. If sum then will use sum of n otherwise will use value entered}

\item{pct_formatting}{should percent formatting be applied}

\item{cum_prop}{if TRUE will add column for cumulative percent}

\item{pct}{if TRUE will add percent column}

\item{cum_sum}{if TRUE will add column for cumulative sum}

\item{head}{returns all values by default}
}
\description{
Add proportion to counts
}
