% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_frech_lo.R
\name{bind_frech_lo}
\alias{bind_frech_lo}
\title{Bind the weighted log odds to a tidy dataset}
\source{
\url{https://doi.org/10.1093/pan/mpn018}
}
\usage{
bind_frech_lo(
  tbl,
  set,
  feature,
  n,
  .alpha = NULL,
  .equation = 15,
  .unweighted = TRUE
)
}
\arguments{
\item{tbl}{A tidy dataset with one row per \code{feature} and \code{set}}

\item{set}{Column of sets between which to compare features, such as
documents for text data}

\item{feature}{Column of features for identifying differences, such as words or
bigrams with text data}

\item{n}{Column containing feature-set counts}

\item{.alpha}{If \code{NULL} uses empirical prior otherwise this is prior for every feature (uninformative)}

\item{.equation}{Whether to compare to dataset (equation 15) or other sets (equation 16)}

\item{.unweighted}{Whether or not to return the unweighted log odds,
in addition to the weighted log odds. Defaults to \code{TRUE}.}
}
\description{
Calculate and bind posterior log odds ratios, assuming a
multinomial model with a Dirichlet prior. The Dirichlet prior
parameters are set using an empirical Bayes approach by default,
but an uninformative prior is also available. Assumes that data
is in a tidy format, and adds the weighted log odds ratio
as a column. Supports non-standard evaluation through the
tidyeval framework.
}
\details{
The arguments \code{set}, \code{feature}, and \code{n}
are passed by expression and support \link[rlang]{quasiquotation};
you can unquote strings and symbols. Grouping is preserved but ignored.

The dataset must have exactly one row per set-feature combination for
this calculation to succeed. Read Monroe, Colaresi, and Quinn (2017) for
more on the weighted log odds ratio.
}
