% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bully.R
\name{bully}
\alias{bully}
\title{Average observations by group}
\usage{
bully(df, col, ..., name = "mean", na.rm = TRUE, sort = TRUE)
}
\arguments{
\item{df}{a tibble to count}

\item{col}{column to average}

\item{...}{variables to group by}

\item{name}{The output column name. If omitted, it will be mean.}

\item{na.rm}{Whether to remove NA values}

\item{sort}{if TRUE will sort in descending order of mean}
}
\description{
Average observations by group
}
