% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_poisson.R
\name{compare_counts}
\alias{compare_counts}
\title{Calculate chance a > b based on closed gamma-poisson form}
\usage{
compare_counts(shape1, rate1, shape2, rate2)
}
\arguments{
\item{shape1}{Shape of 1st vector}

\item{rate1}{Rate of 1st vector}

\item{shape2}{Shape of 2nd vector}

\item{rate2}{Rate of 2nd vector}
}
\description{
Calculate chance a > b based on closed gamma-poisson form
}
