% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_normal.R
\name{compare_normal}
\alias{compare_normal}
\title{Calculate chance a > b based on closed normal form}
\usage{
compare_normal(mean1, sd1, mean2, sd2, log = FALSE)
}
\arguments{
\item{mean1}{Mean of 1st vector}

\item{sd1}{SD of 1st vector}

\item{mean2}{Mean of 2nd vector}

\item{sd2}{SD of 2nd vector}

\item{log}{Whether to use log values}
}
\description{
Calculate chance a > b based on closed normal form
}
