% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_decimal.R
\name{convert_decimal}
\alias{convert_decimal}
\title{Convert decimal to various options}
\usage{
convert_decimal(
  x,
  output = c("all", "ratio", "percent", "decimal", "fraction", "odds"),
  top = 1
)
}
\arguments{
\item{x}{Decimal to convert}

\item{output}{Desired output}

\item{top}{Number of results to return}
}
\description{
Convert decimal to various options
}
