% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_chart.R
\name{dot_chart}
\alias{dot_chart}
\title{Create dot or lollipop chart}
\usage{
dot_chart(df, key, n, top = 25, facet = NULL, lollipop = FALSE)
}
\arguments{
\item{df}{Tibble}

\item{key}{Feature to include on y-axis}

\item{n}{Count to include on x-axis}

\item{top}{How many keys to include}

\item{facet}{(Optional) Feature to facet by}

\item{lollipop}{Create add line to create lollipop chart}
}
\description{
Create dot or lollipop chart
}
