% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_poisson.R
\name{estimate_counts}
\alias{estimate_counts}
\title{Calculate gamma-poisson posterior}
\usage{
estimate_counts(prior_shape, prior_rate, data_count, data_size, plot = F)
}
\arguments{
\item{prior_shape}{Shape set by prior}

\item{prior_rate}{Rate set by prior}

\item{data_count}{Count of collected data}

\item{data_size}{Size of data}

\item{plot}{Choose to list posterior parameters or plot output}
}
\description{
Calculate gamma-poisson posterior
}
