% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_normal.R
\name{estimate_normal}
\alias{estimate_normal}
\title{Calculate normal-normal posterior}
\usage{
estimate_normal(prior_mean, prior_sd, data_mean, data_sd, data_n, plot = F)
}
\arguments{
\item{prior_mean}{Mean set by prior}

\item{prior_sd}{SD set by prior}

\item{data_mean}{Mean of collected data}

\item{data_sd}{SD of collected data}

\item{data_n}{Number of records for collected data}

\item{plot}{Choose to list posterior parameters or plot output}
}
\description{
Calculate normal-normal posterior
}
