% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcount.R
\name{fcount}
\alias{fcount}
\title{Generate counts on data}
\usage{
fcount(df, ..., sort = TRUE, pct = TRUE, round = FALSE,
  denom = "sum", pct_formatting = TRUE, cum_sum = FALSE,
  cum_pct = FALSE, head = NULL)
}
\arguments{
\item{df}{a tibble to count}

\item{...}{variables to group by}

\item{sort}{if TRUE will sort output in descending order of n}

\item{pct}{if TRUE will add percent column}

\item{round}{if TRUE will round to nearest percent}

\item{denom}{denominator for calculating percent. If sum then will use sum of n otherwise will use value entered}

\item{pct_formatting}{should percent formatting be applied}

\item{cum_sum}{if TRUE will add column for cumulative sum}

\item{cum_pct}{if TRUE will add column for cumulative percent}

\item{head}{returns all values by default}
}
\description{
Generate counts on data
}
