% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trend.R
\name{geom_trend}
\alias{geom_trend}
\title{Trend chart}
\usage{
geom_trend(
  tbl,
  .date,
  .n,
  ...,
  percent = FALSE,
  alpha = 0.2,
  se = FALSE,
  x_lab = NULL,
  y_lab = "Count",
  title = NULL,
  facet = NULL,
  scales = "free_y"
)
}
\arguments{
\item{tbl}{A tidy dataset with one count per date per group}

\item{.date}{Date column}

\item{.n}{Count or Percent column}

\item{...}{additional data to pass to ggplot}

\item{percent}{Whether to use percent labels on y-axis}

\item{alpha}{geom_point alpha}

\item{se}{geom_smooth error bars}

\item{x_lab}{label of x-axis}

\item{y_lab}{label of y-axis}

\item{title}{title of plot}

\item{facet}{group for faceting}

\item{scales}{scales passed to facet_wrap}
}
\description{
Create trend chart with points and smoothed line
}
