% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_and_z.R
\name{p_to_z}
\alias{p_to_z}
\title{Calculate gamma-poisson posterior}
\usage{
p_to_z(p, neg_estimate = F, two_tailed = T)
}
\arguments{
\item{p}{Vector of p-values}

\item{neg_estimate}{Is estimated value negative?}

\item{two_tailed}{Was two-tailed test used?}
}
\description{
Calculate gamma-poisson posterior
}
