% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{power_analysis_binomial}
\alias{power_analysis_binomial}
\title{Calculate minimum sample size needed for binomial experiment}
\source{
\url{https://bookdown.org/content/3686/goals-power-and-sample-size.html#computing-power-and-sample-size/}
}
\usage{
power_analysis_binomial(
  gen_prior_mode,
  gen_prior_n,
  hdi_max_width = NULL,
  null_value = NULL,
  rope = c(max(0, null_value - 0.02), min(1, null_value + 0.02)),
  desired_power = 0.8,
  aud_prior_mode = 0.5,
  aud_prior_n = 2,
  hdi_mass = 0.95,
  init_samp_size = 20,
  verbose = TRUE
)
}
\arguments{
\item{gen_prior_mode}{Expected mode of prior}

\item{gen_prior_n}{N of prior (higher leads to more precision)}

\item{hdi_max_width}{Maximum width of HDI}

\item{null_value}{Null value}

\item{rope}{Region of practical equivalence around parameter}

\item{desired_power}{Level of power}

\item{aud_prior_mode}{Prior understanding of mode from audience}

\item{hdi_mass}{Range of HDI}

\item{init_samp_size}{Initial sample size}

\item{verbose}{Should each attempt be displayed?}

\item{aud_prior}{Prior understanding of variation from audience}
}
\description{
Either \code{null_value} or \code{hdi_max_width} must be null,
depending whether the goal is reject a particular null value
or estimating parameter with specified precision
}
\details{
Function was derived by Solomon Kurz's bookdown
converting Kruschke's book into \link{brms} and \link{tidyverse}
}
