% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bigram_stopwords.R
\name{remove_bigram_stopwords}
\alias{remove_bigram_stopwords}
\title{Remove stop words from bigrams}
\usage{
remove_bigram_stopwords(df, bigrams, char_only = TRUE)
}
\arguments{
\item{df}{A tibble containing bigrams}

\item{char_only}{Remove numbers}

\item{bigram}{Column of bigrams}
}
\description{
Remove stop words from bigrams
}
