% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_filter.R
\name{str_filter}
\alias{str_filter}
\title{Filter based on selected text}
\usage{
str_filter(.data, string, pattern)
}
\arguments{
\item{.data}{A tibble, or an object that can be coerced into a tibble.}

\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{pattern}{Pattern to look for.}
}
\description{
Filter based on selected text
}
