% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_predicted_draws.R
\name{summarize_predicted_draws}
\alias{summarize_predicted_draws}
\title{Summarize draws from Stan model}
\usage{
summarize_predicted_draws(predicted_draws, truth, prediction, robust = F)
}
\arguments{
\item{predicted_draws}{Tidy output from add_predicted_draws or predicted_draws}

\item{truth}{Column indicating observed values}

\item{prediction}{Column indicating predicted values}

\item{robust}{Whether to use mean or median}
}
\description{
Summarize draws from Stan model
}
