% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_green.R
\name{theme_green}
\alias{theme_green}
\title{Generate counts on data}
\usage{
theme_green(base_size = 11, base_family = "Century Gothic")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font}
}
\description{
Generate counts on data
}
