% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_group.R
\name{top_n_group}
\alias{top_n_group}
\title{Select top (or bottom) n rows for each group (by value)}
\usage{
top_n_group(df, n, wt, ...)
}
\arguments{
\item{df}{A tibble to filter}

\item{n}{number of rows to return for each group}

\item{wt}{(Optional). The variable to use for ordering. If not specified, defaults to the last variable in the tbl.}

\item{...}{Optional columns to group by}
}
\description{
Select top (or bottom) n rows for each group (by value)
}
