% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_normal.R
\name{update_normal}
\alias{update_normal}
\alias{normal_update}
\title{Calculate normal-normal posterior}
\usage{
normal_update(prior_mean, prior_sd, data_mean, data_sd, data_n, plot = F)
}
\arguments{
\item{prior_mean}{Mean set by prior}

\item{prior_sd}{SD set by prior}

\item{data_mean}{Mean of collected data}

\item{data_se}{SD of data divided by square root of observations}
}
\description{
Calculate normal-normal posterior
}
