% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_and_z.R
\name{z_to_p}
\alias{z_to_p}
\title{Calculate gamma-poisson posterior}
\usage{
z_to_p(z, two_tailed = T)
}
\arguments{
\item{z}{Vector of z-values}

\item{two_tailed}{Was two-tailed test used?}
}
\description{
Calculate gamma-poisson posterior
}
