% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blow.R
\name{add_blow}
\alias{add_blow}
\title{Bind the weighted log odds to a tidy dataset}
\source{
\url{https://doi.org/10.1093/pan/mpn018}
}
\usage{
add_blow(
  tbl,
  group,
  feature,
  n,
  topic = NULL,
  .alpha = NULL,
  .compare = c("all", "groups"),
  .unweighted = TRUE,
  .variance = TRUE,
  .odds = TRUE,
  .prob = TRUE
)
}
\arguments{
\item{tbl}{A tidy dataset with one row per feature and set}

\item{group}{Column of groups between which to compare features, such as
documents for text data}

\item{feature}{Column of features for identifying differences, such as words or
bigrams with text data}

\item{n}{Column containing feature-set counts}

\item{topic}{(Optional) topic to compare groups within}

\item{.alpha}{(Optional) Prior frequency of each feature. Can be null to
calculate empirically or uninformative prior set for all features}

\item{.compare}{Whether to compare group-feature combination to dataset or
all other groups}

\item{.unweighted}{Whether to include point estimate log odds}

\item{.variance}{Whether to include variance of feature}

\item{.odds}{Whether to include odds of seeing feature within group}

\item{.prob}{Whether to include probability for feature within group}
}
\description{
Calculate and bind the log odds ratio, weighted by an uninformative Dirichlet
prior, of a tidy dataset to the dataset itself. The weighted log odds ratio
is added as a column. This functions supports non-standard evaluation through
the tidyeval framework.
}
\details{
The arguments \code{group}, \code{feature}, \code{n}, and \code{topic}
are passed by expression and support \link[rlang]{quasiquotation};
you can unquote strings and symbols. Grouping is preserved but ignored.

The dataset must have exactly one row per group-feature combination for
this calculation to succeed. Read Monroe, Colaresi, and Quinn (2017) for
more on the weighted log odds ratio.
}
