% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsummary.R
\name{fsummary}
\alias{fsummary}
\title{Generate summary statistics on data}
\usage{
fsummary(df, n, ..., sort = T)
}
\arguments{
\item{df}{A tibble to summarize}

\item{n}{Column to summarize}

\item{...}{Optional columns to group by}

\item{sort}{if TRUE will sort output in descending order of n}
}
\description{
Generate summary statistics on data
}
